backend/requirements.txt
uvicorn==0.29.0
fastapi==0.110.1
beanie==1.25.0
motor==3.4.0
python-dotenv==1.0.1